Fix installing extra packages on linux/386.
slackpkg will not install anything since `uname -a` will return x86_64
when running on a 64bit host. I have thought about configuring slackpkg
with an explicit ARCH in the base images so this just works. I am not
sure if that is the best solution - for now, I am going to fix it here.
(I also already manually fix it in the "full" images - so maybe I should
just fix the base).